+++ engine/src/test/java/net/jqwik/api/ArbitrariesTests.java
@@ -78,6 +78,15 @@ class ArbitrariesTests {
 	}
 
 	@Example
+	void ofValueSet() {
+		Set<String> valueSet = new HashSet<>(Arrays.asList("1", "hallo", "test"));
+		Arbitrary<String> stringArbitrary = Arbitraries.of(valueSet);
+		RandomGenerator<String> generator = stringArbitrary.generator(1);
+		assertAllGenerated(generator, (String value) -> Arrays.asList("1", "hallo", "test").contains(value));
+		assertAtLeastOneGeneratedOf(generator, "1", "hallo", "test");
+	}
+
+	@Example
 	void ofSuppliers() {
 		Arbitrary<List<String>> listArbitrary = Arbitraries.ofSuppliers(ArrayList::new, ArrayList::new);
 		RandomGenerator<List<String>> generator = listArbitrary.generator(1);
@@ -92,6 +101,19 @@ class ArbitrariesTests {
 		@SuppressWarnings("unchecked")
 		Supplier<List<String>>[] suppliers = new Supplier[]{ArrayList::new, ArrayList::new};
 		List<Supplier<List<String>>> supplierList = Arrays.asList(suppliers);
+		Arbitrary<List<String>> listArbitrary = Arbitraries.ofSuppliers(supplierList);
+		RandomGenerator<List<String>> generator = listArbitrary.generator(1);
+		assertAllGenerated(generator, (List<String> value) -> {
+			assertThat(value).isEmpty();
+			value.add("aString");
+		});
+	}
+
+	@Example
+	void ofSupplierSet() {
+		@SuppressWarnings("unchecked")
+		Supplier<List<String>>[] suppliers = new Supplier[]{ArrayList::new, ArrayList::new};
+		Set<Supplier<List<String>>> supplierList = new HashSet<>(Arrays.asList(suppliers));
 		Arbitrary<List<String>> listArbitrary = Arbitraries.ofSuppliers(supplierList);
 		RandomGenerator<List<String>> generator = listArbitrary.generator(1);
 		assertAllGenerated(generator, (List<String> value) -> {